Move GetNoUsedFds to testutil
